IRender接口调整：简化DrawRect接口，删除冗余的接口函数，利于后期绘制引擎维护。

+++ duilib/Control/ListCtrlReportView.cpp
@@ -763,7 +763,7 @@ void ListCtrlReportView::PaintGridLines(IRender* pRender)
 
             UiPointF pt1(viewRect.left, yPos);
             UiPointF pt2(viewRect.right, yPos);
-            pRender->DrawLine(pt1, pt2, rowLineColor, nRowLineWidth);
+            pRender->DrawLine(pt1, pt2, rowLineColor, (float)nRowLineWidth);
         }
     }
 }
